:recycle: Add: router with middleware
