+++ backend/cmd/api/main.go
@@ -1,38 +1,20 @@
 package main
 
 import (
-	"errors"
-	"net/http"
-
 	"github.com/labstack/echo/v4"
 	"github.com/naohito-T/tinyurl/backend/internal/rest/middleware"
+	"github.com/naohito-T/tinyurl/backend/internal/rest/router"
 )
 
 const (
-	defaultPort = "6500"
+	defaultPort = ":6500"
 )
 
 func main() {
 	// Echo instance
 	e := echo.New()
 	middleware.CustomMiddleware(e)
-	// Routes
-	e.GET("/", hello)
+	router.NewRouter(e)
 	// e.Startでエラーが発生した場合、Fatalでプログラムを終了する
 	e.Logger.Fatal(e.Start(defaultPort))
-}
-
-func hello(c echo.Context) error {
-	println("Hello, World!")
-	if err := isValid("hello"); err != nil {
-		return err
-	}
-	return c.String(http.StatusOK, "Hello, World!")
-}
-
-func isValid(txt string) error {
-	if txt == "" {
-		return errors.New("Invalid")
-	}
-	return nil
 }
